+++ app/scripts/services/languages/language.provider.js
@@ -77,6 +77,17 @@
             return this;
         };
 
+        this.logsEnabled = function ($value) {
+            if (Utils.isBoolean($value)) {
+                config.debug.logs.language.enabled = $value;
+            }
+            else {
+                Logs.errorInProvider(data.provider);
+                Logs.dataMustBeBoolean('logsEnabled');
+            }
+            return this;
+        };
+
         this.$get = language;
 
         language.$inject = [];